Fix deprecation warning about syscall on macOS
Before, the compiler said, "warning: 'syscall' is deprecated: first deprecated
in macOS 10.12 - syscall(2) is unsupported; please switch to a supported
interface. …" To eliminate the warning, this commit changes the call from
`syscall()` to `pthread_threadid_np()` on Apple systems to get the thread ID,
based on [this StackOverflow answer][1]. Note that the type of the thread ID is
different from that on Linux, so this commit also adds a separate
`thread_id_apple` field to the `ThreadHealth` Protobuf message.

[1]: https://stackoverflow.com/a/65494514