Add support for Iver Remote Helm version 5
This adds a field to IverConfig to specify the remote helm major version and
makes the changes necessary for the version >= 5 case. This change is basically
a copy of the following commit for Goby2 by @tsaubergine:
https://github.com/GobySoft/goby/commit/430965ed0306d061a6059ee0762588ac735dcde0
Thanks also to @tsaubergine for identifying that the interface change occurred
with Remote Helm version 5.

Co-authored-by: Toby Schneider <toby@gobysoft.org>