Add config to assign Iver modes to frontseat states
The appropriate assignments of Iver mission modes to frontseat states (i.e.
which mission modes should accept backseat commands) depend on how the user
plans to operate the vehicle. For example, while the default assignments are
useful in many cases, in some cases it may be more convenient to keep the
frontseat in control when in NORMAL mode (to allow the frontseat to drive the
vehicle to a waypoint and park) and allow backseat control when in PARKING
mode (to allow the backseat to take over once the vehicle parks at the
waypoint).